+++ Sources/SheetMusicMSCX/Decoders/MSCXDecoder+Clef.swift
@@ -4,7 +4,7 @@ import SheetMusicXMLTools
 
 extension Clef {
     private static let consumedClefChildren: Set = [
-        "color", "concertClefType", "transposingClefType", "visible",
+        "color", "concertClefType", "offset", "transposingClefType", "visible",
     ]
 
     static func decode(_ node: XMLTreeNode) throws -> Clef {