+++ Sources/SheetMusicMSCX/Encoders/MSCXEncoder+BarLine.swift
@@ -13,6 +13,7 @@ extension BarLine {
         }
         children.append(contentsOf: elementProperties.mscxChildren())
         appendPreservedMarkup(preservedMarkup, to: &children, options: options)
+        children += elementProperties.mscxTrailingChildren()
         return XMLTreeNode(name: "BarLine", children: children)
     }
 }